[mesh] Allow strings instead of arrays in setXXXArrays() [mesh] Use 32bit index array nstead of 16bit

+++ 2dsg/gmesh.h
@@ -19,12 +19,12 @@ public:
     bool is3d();
 
     void setVertex(int i, float x, float y, float z);
-    void setIndex(int i, unsigned short index);
+    void setIndex(int i, unsigned int index);
     void setColor(int i, unsigned int color, float alpha);
     void setTextureCoordinate(int i, float u, float v);
 
     void setVertexArray(const float *vertices, size_t size);
-    void setIndexArray(const unsigned short *indices, size_t size);
+    void setIndexArray(const unsigned int *indices, size_t size);
     void setColorArray(const unsigned int *colors, const float *alphas, size_t size);
     void setTextureCoordinateArray(const float *textureCoordinates, size_t size);
     void setGenericArray(int index,const void *pointer, ShaderProgram::DataType type, int mult, int count);
@@ -45,7 +45,7 @@ public:
     size_t getTextureCoordinateArraySize() const;
 
     void getVertex(int i, float *x, float *y, float *z) const;
-    void getIndex(int i, unsigned short *index) const;
+    void getIndex(int i, unsigned int *index) const;
     void getColor(int i, unsigned int *color, float *alpha) const;
     void getTextureCoordinate(int i, float *u, float *v) const;
 
@@ -66,7 +66,7 @@ private:
         float alpha;
     };
     VertexBuffer<float> vertices_;
-    VertexBuffer<unsigned short> indices_;
+    VertexBuffer<unsigned int> indices_;
     VertexBuffer<unsigned char> colors_;
     std::vector<Color> originalColors_;
     VertexBuffer<float> textureCoordinates_;
